Support for stack outputs
Adds support for top-level exports in the main script of a Pulumi Program to be captured as stack-level output properties.

This is currently accomplised by alwasy creating a special ComponentResource of type `pulumi:pulumi:Stack` whose output properties are the exports of the program.

This will in general be displayed last in a preview/update operation, providing a summary of the most important values indicated by the application author.